+++ NEWS.md
@@ -4,7 +4,11 @@
 Due to using several APIs only recently added, 0.2.0 onwards requires current
 `master` branch of botan (upcoming 2.8, to be released in early October).
 
-Adds scrypt. Adds various util functions to existing types
+- Adds scrypt.
+- Add NIST key wrapping.
+- Adds various util functions to existing types.
+- Fix a bug affecting DH/ECDH (it always returned exactly 128 bytes,
+  instead of taking the requested KDF output length from the caller).
 
 ## 0.1.5 2018-8-5