+++ NEWS.md
@@ -1,5 +1,5 @@
 
-## 0.2.0 Not Yet Released
+## 0.2.0 2018-8-15
 
 Due to using several APIs only recently added, 0.2.0 onwards requires current
 `master` branch of botan (upcoming 2.8, to be released in early October).